fix: unify daily memory startup and summary rules
