+++ src/auto-reply/reply/startup-context.ts
@@ -3,11 +3,7 @@ import path from "node:path";
 import { resolveUserTimezone } from "../../agents/date-time.js";
 import type { OpenClawConfig } from "../../config/config.js";
 import { openBoundaryFile } from "../../infra/boundary-file-read.js";
-import {
-  isSessionSummaryDailyMemory,
-  listRecentDailyMemoryFiles,
-  type DailyMemoryFileEntry,
-} from "../../memory-host-sdk/runtime-files.js";
+import { selectStartupDailyMemoryPaths } from "../../memory-host-sdk/host/daily-startup-selection.js";
 
 const STARTUP_MEMORY_FILE_MAX_BYTES = 16_384;
 const STARTUP_MEMORY_FILE_MAX_CHARS = 1_200;
@@ -67,113 +63,6 @@ function resolveStartupContextLimits(cfg?: OpenClawConfig) {
   };
 }
 
-function formatDateStamp(nowMs: number, timezone: string): string {
-  const parts = new Intl.DateTimeFormat("en-US", {
-    timeZone: timezone,
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  }).formatToParts(new Date(nowMs));
-  const year = parts.find((part) => part.type === "year")?.value;
-  const month = parts.find((part) => part.type === "month")?.value;
-  const day = parts.find((part) => part.type === "day")?.value;
-  if (year && month && day) {
-    return `${year}-${month}-${day}`;
-  }
-  return new Date(nowMs).toISOString().slice(0, 10);
-}
-
-function shiftDateStampByCalendarDays(stamp: string, offsetDays: number): string {
-  const [yearRaw, monthRaw, dayRaw] = stamp.split("-").map((part) => Number.parseInt(part, 10));
-  if (!yearRaw || !monthRaw || !dayRaw) {
-    return stamp;
-  }
-  const shifted = new Date(Date.UTC(yearRaw, monthRaw - 1, dayRaw + offsetDays));
-  return shifted.toISOString().slice(0, 10);
-}
-
-function buildStartupLocalDayCandidates(params: {
-  nowMs: number;
-  timezone: string;
-  dailyMemoryDays: number;
-}): string[] {
-  const orderedDays: string[] = [];
-  const localToday = formatDateStamp(params.nowMs, params.timezone);
-  const daysToScan = Math.min(
-    STARTUP_MEMORY_DAILY_DAYS_CAP,
-    Math.max(1, Math.trunc(params.dailyMemoryDays)),
-  );
-  for (let offset = 0; offset < daysToScan; offset += 1) {
-    orderedDays.push(shiftDateStampByCalendarDays(localToday, -offset));
-  }
-  return orderedDays;
-}
-
-function groupStartupPathsByDay(params: {
-  entries: DailyMemoryFileEntry[];
-  prioritizedPaths?: ReadonlySet<string>;
-}): Map<string, string[]> {
-  const prioritizedPaths = params.prioritizedPaths ?? new Set<string>();
-  const entriesByDay = new Map<string, Array<{ entry: DailyMemoryFileEntry; index: number }>>();
-  for (const [index, entry] of params.entries.entries()) {
-    const dayEntries = entriesByDay.get(entry.day);
-    if (dayEntries) {
-      dayEntries.push({ entry, index });
-      continue;
-    }
-    entriesByDay.set(entry.day, [{ entry, index }]);
-  }
-  return new Map(
-    [...entriesByDay.entries()].map(([day, entries]) => [
-      day,
-      entries
-        .toSorted((left, right) => {
-          const leftPriority = prioritizedPaths.has(left.entry.relativePath)
-            ? 0
-            : left.entry.canonical
-              ? 1
-              : 2;
-          const rightPriority = prioritizedPaths.has(right.entry.relativePath)
-            ? 0
-            : right.entry.canonical
-              ? 1
-              : 2;
-          if (leftPriority !== rightPriority) {
-            return leftPriority - rightPriority;
-          }
-          if (leftPriority === 0 && left.entry.mtimeMs !== right.entry.mtimeMs) {
-            return right.entry.mtimeMs - left.entry.mtimeMs;
-          }
-          return left.index - right.index;
-        })
-        .map(({ entry }) => entry.relativePath),
-    ]),
-  );
-}
-
-function interleaveStartupPathsByDay(params: {
-  orderedDays: string[];
-  pathsByDay: Map<string, string[]>;
-}): string[] {
-  const relativePaths: string[] = [];
-  for (let offset = 0; params.orderedDays.length > 0; offset += 1) {
-    let addedAtOffset = false;
-    for (const day of params.orderedDays) {
-      const dayPaths = params.pathsByDay.get(day);
-      const nextPath = dayPaths?.[offset];
-      if (!nextPath) {
-        continue;
-      }
-      relativePaths.push(nextPath);
-      addedAtOffset = true;
-    }
-    if (!addedAtOffset) {
-      break;
-    }
-  }
-  return relativePaths;
-}
-
 async function resolveStartupDailyMemoryPaths(params: {
   workspaceDir: string;
   dailyMemoryDays: number;
@@ -182,140 +71,19 @@ async function resolveStartupDailyMemoryPaths(params: {
   maxFileBytes: number;
   readCache: Map<string, string | null>;
 }): Promise<string[]> {
-  const localDayCandidates = buildStartupLocalDayCandidates({
+  return await selectStartupDailyMemoryPaths({
+    workspaceDir: params.workspaceDir,
+    dailyMemoryDays: params.dailyMemoryDays,
     nowMs: params.nowMs,
     timezone: params.timezone,
-    dailyMemoryDays: params.dailyMemoryDays,
-  });
-  const localToday = formatDateStamp(params.nowMs, params.timezone);
-  const localYesterday = shiftDateStampByCalendarDays(localToday, -1);
-  const utcToday = formatDateStamp(params.nowMs, "UTC");
-  const targetDays = [...localDayCandidates];
-  if (params.dailyMemoryDays === 1 && !targetDays.includes(localYesterday)) {
-    targetDays.push(localYesterday);
-  }
-  if (utcToday !== localToday && !targetDays.includes(utcToday)) {
-    targetDays.push(utcToday);
-  }
-  const currentLocalDay = localDayCandidates[0];
-  const entries = await listRecentDailyMemoryFiles({
-    memoryDir: path.join(params.workspaceDir, "memory"),
-    days: targetDays,
-    // Startup context should stay read-only even on the first reset/new turn.
-    persistIndex: false,
-  });
-  const summaryPriorityDays = [...new Set([currentLocalDay, localYesterday, utcToday])].filter(
-    (day) => Boolean(day) && targetDays.includes(day),
-  );
-  const sessionSummaryPaths = new Set<string>();
-  for (const day of summaryPriorityDays) {
-    const daySummaryPaths = await resolveStartupSessionSummaryPaths({
-      workspaceDir: params.workspaceDir,
-      entries,
-      day,
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    });
-    for (const relativePath of daySummaryPaths) {
-      sessionSummaryPaths.add(relativePath);
-    }
-  }
-  const pathsByDay = groupStartupPathsByDay({
-    entries,
-    prioritizedPaths: sessionSummaryPaths,
-  });
-  const selectedDays: string[] = [];
-  const existingLocalDays = localDayCandidates.filter(
-    (day) => (pathsByDay.get(day)?.length ?? 0) > 0,
-  );
-  const boundaryDay =
-    utcToday !== localToday &&
-    entries.some((entry) => entry.day === utcToday && sessionSummaryPaths.has(entry.relativePath))
-      ? utcToday
-      : null;
-
-  if (
-    currentLocalDay &&
-    (pathsByDay.get(currentLocalDay)?.length ?? 0) > 0 &&
-    (await hasReadableStartupMemoryPathForDay({
-      workspaceDir: params.workspaceDir,
-      relativePaths: pathsByDay.get(currentLocalDay) ?? [],
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    }))
-  ) {
-    selectedDays.push(currentLocalDay);
-  }
-  if (
-    params.dailyMemoryDays === 1 &&
-    !selectedDays.includes(localYesterday) &&
-    selectedDays.length === 0 &&
-    (await hasReadableStartupMemoryPathForDay({
-      workspaceDir: params.workspaceDir,
-      relativePaths: (pathsByDay.get(localYesterday) ?? []).filter((relativePath) =>
-        sessionSummaryPaths.has(relativePath),
-      ),
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    }))
-  ) {
-    selectedDays.push(localYesterday);
-  }
-  if (
-    boundaryDay &&
-    !selectedDays.includes(boundaryDay) &&
-    (await hasReadableStartupMemoryPathForDay({
-      workspaceDir: params.workspaceDir,
-      relativePaths: pathsByDay.get(boundaryDay) ?? [],
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    }))
-  ) {
-    selectedDays.push(boundaryDay);
-  }
-  for (const day of existingLocalDays) {
-    if (selectedDays.length >= params.dailyMemoryDays) {
-      break;
-    }
-    if (
-      !selectedDays.includes(day) &&
-      (await hasReadableStartupMemoryPathForDay({
+    readDailyMemory: async (relativePath) =>
+      await readStartupMemoryFile({
         workspaceDir: params.workspaceDir,
-        relativePaths: pathsByDay.get(day) ?? [],
+        relativePath,
         maxFileBytes: params.maxFileBytes,
         readCache: params.readCache,
-      }))
-    ) {
-      selectedDays.push(day);
-    }
-  }
-
-  // Read the first file for each selected day before extra same-day variants so
-  // UTC-boundary summaries are not starved by local-day variant fan-out.
-  return interleaveStartupPathsByDay({
-    orderedDays: selectedDays,
-    pathsByDay,
+      }),
   });
-}
-
-async function hasReadableStartupMemoryPathForDay(params: {
-  workspaceDir: string;
-  relativePaths: string[];
-  maxFileBytes: number;
-  readCache: Map<string, string | null>;
-}): Promise<boolean> {
-  for (const relativePath of params.relativePaths) {
-    const content = await readStartupMemoryFile({
-      workspaceDir: params.workspaceDir,
-      relativePath,
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    });
-    if (content?.trim()) {
-      return true;
-    }
-  }
-  return false;
 }
 
 function trimStartupMemoryContent(content: string, maxChars: number): string {
@@ -435,31 +203,6 @@ async function readStartupMemoryFile(params: {
   } finally {
     await closeFd(opened.fd);
   }
-}
-
-async function resolveStartupSessionSummaryPaths(params: {
-  workspaceDir: string;
-  entries: DailyMemoryFileEntry[];
-  day?: string;
-  maxFileBytes: number;
-  readCache?: Map<string, string | null>;
-}): Promise<Set<string>> {
-  const sessionSummaryPaths = new Set<string>();
-  for (const entry of params.entries) {
-    if (params.day && entry.day !== params.day) {
-      continue;
-    }
-    const content = await readStartupMemoryFile({
-      workspaceDir: params.workspaceDir,
-      relativePath: entry.relativePath,
-      maxFileBytes: params.maxFileBytes,
-      readCache: params.readCache,
-    });
-    if (content && isSessionSummaryDailyMemory(content)) {
-      sessionSummaryPaths.add(entry.relativePath);
-    }
-  }
-  return sessionSummaryPaths;
 }
 
 export async function buildSessionStartupContextPrelude(params: {
